+++ texk/web2c/luatexdir/luamd5/md5.c
@@ -6,13 +6,14 @@
 
 
 #include <string.h>
+#include <stdint.h>
 
-#include "md5.h"
+#include "luamd5.h"
 
 
 #define WORD 32
 #define MASK 0xFFFFFFFF
-typedef unsigned long WORD32;
+typedef uint32_t WORD32;
 
 
 /**
